requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
Flask_SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
alembic==1.13.2
passlib[bcrypt]==1.7.4
requests==2.32.3
pydantic==2.9.2
PyMySQL==1.1.1
google-genai>=1.0.0
Babel==2.12.1
Flask-Babel==4.0.0